version 0.5 on the way
Made lots of changes to the current build like making a chapter class and editing the description and readme class. Lots of functionality works but there's a bug when loading the file where it would successfully pull up but it would not load the contents. have to fix that soon